+++ modules/utils/bindApp.js
@@ -1,7 +1,5 @@
 /* jshint -W058 */
-
-var Connection = require('../Connection');
-var Location = require('../Location');
+var createConnection = require('./createConnection');
 
 /**
  * HTTP status codes that don't have entities.
@@ -14,86 +12,6 @@ var STATUS_WITHOUT_CONTENT = {
 };
 
 /**
- * Standard ports for HTTP protocols.
- */
-var STANDARD_PORTS = {
-  'http:': 80,
-  'https:': 443
-};
-
-function getHost(nodeServer) {
-  if (process.env.SERVER_NAME)
-    return process.env.SERVER_NAME;
-
-  var address = nodeServer.address();
-
-  if (typeof address === 'string')
-    return address;
-
-  return address.address + ':' + address.port;
-}
-
-/**
- * Creates a new Location object that is reverse-proxy aware.
- */
-function createLocation(nodeRequest, nodeServer) {
-  var headers = nodeRequest.headers;
-
-  var protocol;
-  if (process.env.HTTPS === 'on') {
-    protocol = 'https:';
-  } else if (headers['x-forwarded-ssl'] === 'on') {
-    protocol = 'https:';
-  } else if (headers['x-forwarded-scheme']) {
-    protocol = headers['x-forwarded-scheme'];
-  } else if (headers['x-forwarded-proto']) {
-    protocol = headers['x-forwarded-proto'].split(',')[0];
-  } else {
-    protocol = nodeRequest.protocol;
-  }
-
-  var host;
-  if (headers['x-forwarded-host']) {
-    var hosts = headers['x-forwarded-host'].split(/,\s?/);
-    host = hosts[hosts.length - 1];
-  } else {
-    host = headers['host'] || getHost(nodeServer);
-  }
-
-  var hostParts = host.split(':', 2);
-  var hostname = hostParts[0];
-  var port = hostParts[1] || headers['x-forwarded-port'];
-
-  if (port == null) {
-    if (headers['x-forwarded-host']) {
-      port = STANDARD_PORTS[protocol];
-    } else if (headers['x-forwarded-proto']) {
-      port = STANDARD_PORTS[headers['x-forwarded-proto'].split(',')[0]];
-    }
-  }
-
-  var path = nodeRequest.url;
-  var index = path.indexOf('?');
-
-  var pathname, search;
-  if (index !== -1) {
-    pathname = path.substring(0, index);
-    search = path.substring(index);
-  } else {
-    pathname = path;
-    search = '';
-  }
-
-  return new Location({
-    protocol: protocol,
-    hostname: hostname,
-    port: port,
-    pathname: pathname,
-    search: search
-  });
-}
-
-/**
  * Binds the given app to the "request" event of the given node HTTP server
  * so that it is called whenever the server receives a new request.
  *
@@ -101,19 +19,7 @@ function createLocation(nodeRequest, nodeServer) {
  */
 function bindApp(app, nodeServer) {
   function requestHandler(nodeRequest, nodeResponse) {
-    var conn = new Connection({
-      version: nodeRequest.httpVersion,
-      method: nodeRequest.method,
-      location: createLocation(nodeRequest, nodeServer),
-      headers: nodeRequest.headers,
-      content: nodeRequest,
-      remoteHost: nodeRequest.connection.remoteAddress,
-      remotePort: nodeRequest.connection.remotePort
-    });
-
-    nodeRequest.on('close', function () {
-      conn.onClose();
-    });
+    var conn = createConnection(nodeRequest);
 
     conn.call(app).then(function () {
       var isHead = conn.method === 'HEAD';